@@ -9,7 +9,7 @@ packages = [{ include = "app" }]
 
 
 [tool.poetry.dependencies]
-python = "^3.10"
+python = "^3.12"
 fastapi = "^0.109.0"
 uvicorn = {extras = ["standard"], version = "^0.27.0.post1"}
 redis = {extras = ["hiredis"], version = "^5.0.1"}
